+++ src/Steps/Login.vue
@@ -1,35 +1,21 @@
 <template>
   <v-container>
-    <v-layout
-      row
-      wrap
-    >
-      <v-flex v-if="notLoggedIn">
+    <v-layout>
+      <v-flex>
+        <SetBackendForm
+          v-if="setBackend"
+          @setBackend="toggleSetBackend"
+        />
         <LoginForm
-          v-if="!createAccount"
+          v-else-if="!createAccount"
           @createAccount="toggleCreateAccount"
+          @setBackend="toggleSetBackend"
         />
         <CreateUserForm
           v-else
           @login="toggleCreateAccount"
+          @setBackend="toggleSetBackend"
         />
-      </v-flex>
-      <v-flex v-else>
-        <v-card>
-          <v-card-text>
-            <h1 style="text-align: center">
-              Hello, {{ auth.user.login }}
-            </h1>
-            <p>
-              Click "continue" or <v-btn
-                color="primary"
-                @click="logout"
-              >
-                Logout
-              </v-btn>
-            </p>
-          </v-card-text>
-        </v-card>
       </v-flex>
     </v-layout>
   </v-container>
@@ -45,61 +31,28 @@
 import _ from 'lodash';
 import LoginForm from '../Components/Authentication/LoginForm.vue';
 import CreateUserForm from '../Components/Authentication/CreateUserForm.vue';
+import SetBackendForm from '../Components/Authentication/SetBackendForm.vue';
 
 export default {
   name: 'Login',
 
   components: {
     LoginForm,
-    CreateUserForm
+    CreateUserForm,
+    SetBackendForm,
   },
 
   data: () => ({
-    username: '',
-    password: '',
-    error: '',
-    createAccount: false
+    createAccount: false,
+    setBackend: false,
   }),
 
-  computed: {
-    /**
-     * only let the user continue if they are authenticated
-     */
-    readyToContinue() {
-      if (!_.isEmpty(this.auth)) {
-        return true;
-      }
-      return false;
-    },
-    /**
-     * a boolean value for whether or not the user is logged in
-     */
-    notLoggedIn() {
-      return _.isEmpty(this.auth);
-    },
-    /**
-     * return the auth value from the store.
-     */
-    auth() {
-      return this.$store.state.auth;
-    }
-  },
-
   methods: {
-    /**
-     * nothing to do when continue is pressed
-     */
-    continueAction() {
-
-    },
     toggleCreateAccount() {
       this.createAccount = !this.createAccount;
     },
-    /**
-     * clear the store on logout.
-     */
-    logout() {
-      this.$store.commit('setAuth', {});
+    toggleSetBackend() {
+      this.setBackend = !this.setBackend;
     }
   },