+++ src/Steps/SetReviewers.vue
@@ -1,5 +1,0 @@
-<template>
-  <div>
-    <h1>REviewers</h1>
-  </div>
-</template>