+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.40] = 2020-4-12
+### Changed
+- add ```vuex-persistedstate``` to store state on refresh
+- spread 'steps' across routes with ```vue-router```
+    - add authentication rules for each route
+- add app bar to entire app, minor other UI changes
+
 ## [0.1.39] = 2020-4-9
 ### Changed
 - protocol-builder 0.1.18 -> 0.1.20